frontend/requirements.txt
flask==3.0.0
flask-bcrypt==1.0.1
PyJWT==2.8.0
requests==2.32.5
dotenv==0.9.9
pymongo==4.10.1

